offer-16 completed

+++ problemset/offer-16.cpp
@@ -3,7 +3,7 @@
 
 class Solution {
 private:
-    double myPowR(double x, long long n) {
+    double myPowR(double x, int n) {
         if(n == 0) return 1;
 
         if(n < 0) {
